requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=2.6.0
discord.py>=2.6.4